# server/api/openapi.gen.yaml
...
components:
    parameters:
        DocListCursor:
            in: query
            name: cursor
            required: false
            schema:
                type: string
        DocListDateFrom:
            in: query
            name: dateFrom
            required: false
            schema:
                format: date
                type: string
        DocListDateTo:
            in: query
            name: dateTo
            required: false
            schema:
                format: date
                type: string
        DocListLimit:
            in: query
            name: limit
            required: false
            schema:
                default: 50
                type: integer
        DocListQ:
            in: query
            name: q
            required: false
            schema:
                type: string
        DocListStatus:
            in: query
            name: status
            required: false
            schema:
                type: string
        DocListWarehouseId:
            in: query
            name: warehouseId
            required: false
            schema:
                format: uuid
                type: string
        DocumentId:
            in: path
            name: id
            required: true
            schema:
                format: uuid
                type: string
...
    responses:
        Conflict:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Resource already exists
        Forbidden:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Authenticated but not permitted
        NotFound:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Resource not found
...
        Unauthorized:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Not authenticated
        ValidationError:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Request failed validation
    schemas:
...
        DecimalString:
            description: A decimal quantity or money value, string-encoded to avoid float loss
            example: "12.500"
            type: string
...
        DocumentActor:
            properties:
                displayName:
                    type: string
                id:
                    format: uuid
                    type: string
            required:
                - id
                - displayName
            type: object
        DocumentStatus:
            description: draft is editable; posted is immutable; reversed was cancelled by a reversal; cancelled is terminal for PO/SO (ADR-0018)
            enum:
                - draft
                - posted
                - reversed
                - cancelled
            type: string
        Error:
            properties:
                code:
                    type: string
                fields:
                    additionalProperties:
                        type: string
                    description: Per-field validation messages, when applicable
                    type: object
                message:
                    type: string
            required:
                - code
                - message
            type: object
...
        StockTransfer:
            properties:
                createdAt:
                    format: date-time
                    type: string
                createdBy:
                    $ref: '#/components/schemas/DocumentActor'
                docDate:
                    format: date
                    type: string
                docNumber:
                    nullable: true
                    type: string
                fromWarehouseId:
                    format: uuid
                    type: string
                id:
                    format: uuid
                    type: string
                lines:
                    items:
                        $ref: '#/components/schemas/StockTransferLine'
                    type: array
                notes:
                    type: string
                postedAt:
                    format: date-time
                    type: string
                postedBy:
                    $ref: '#/components/schemas/DocumentActor'
                reversedById:
                    format: uuid
                    nullable: true
                    type: string
                reversesId:
                    format: uuid
                    nullable: true
                    type: string
                status:
                    $ref: '#/components/schemas/DocumentStatus'
                toWarehouseId:
                    format: uuid
                    type: string
            required:
                - id
                - status
                - fromWarehouseId
                - toWarehouseId
                - docDate
                - notes
                - lines
                - createdAt
                - createdBy
            type: object
        StockTransferInput:
            properties:
                docDate:
                    format: date
                    type: string
                fromWarehouseId:
                    format: uuid
                    type: string
                lines:
                    items:
                        $ref: '#/components/schemas/StockTransferLineInput'
                    minItems: 1
                    type: array
                notes:
                    type: string
                toWarehouseId:
                    format: uuid
                    type: string
            required:
                - fromWarehouseId
                - toWarehouseId
                - docDate
                - lines
            type: object
        StockTransferLine:
            properties:
                batchId:
                    format: uuid
                    nullable: true
                    type: string
                id:
                    format: uuid
                    type: string
                lineNo:
                    type: integer
                productId:
                    format: uuid
                    type: string
                qty:
                    $ref: '#/components/schemas/DecimalString'
                uom:
                    type: string
            required:
                - id
                - lineNo
                - productId
                - uom
                - qty
            type: object
        StockTransferLineInput:
            properties:
                batchId:
                    format: uuid
                    nullable: true
                    type: string
                productId:
                    format: uuid
                    type: string
                qty:
                    $ref: '#/components/schemas/DecimalString'
                uom:
                    minLength: 1
                    type: string
            required:
                - productId
                - uom
                - qty
            type: object
        StockTransferList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/StockTransfer'
                    type: array
                nextCursor:
                    nullable: true
                    type: string
            required:
                - items
                - nextCursor
            type: object
...
    securitySchemes:
        sessionCookie:
            in: cookie
            name: sinta_session
            type: apiKey
info:
    description: |
        Multi-tenant inventory management for Indonesian SME distributors. Contract-first: this file is the single source of truth for both the Go server (oapi-codegen) and the TypeScript client (openapi-typescript). Conventions: camelCase fields, kebab-case plural paths, /v1 prefix (docs/reference/CONVENTIONS.md).
    title: Sinta API
    version: 0.1.0
openapi: 3.0.3
paths:
...
    /stock-transfers:
        get:
            operationId: listStockTransfers
            parameters:
                - $ref: '#/components/parameters/DocListStatus'
                - $ref: '#/components/parameters/DocListWarehouseId'
                - $ref: '#/components/parameters/DocListDateFrom'
                - $ref: '#/components/parameters/DocListDateTo'
                - $ref: '#/components/parameters/DocListQ'
                - $ref: '#/components/parameters/DocListCursor'
                - $ref: '#/components/parameters/DocListLimit'
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StockTransferList'
                    description: Stock transfers
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
            summary: List stock transfers
            tags:
                - documents
        post:
            operationId: createStockTransfer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StockTransferInput'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StockTransfer'
                    description: Draft created
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "422":
                    $ref: '#/components/responses/ValidationError'
            summary: Create a draft stock transfer
            tags:
                - documents
    /stock-transfers/{id}:
        delete:
            operationId: deleteStockTransfer
            responses:
                "204":
                    description: Deleted
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
            summary: Delete a draft stock transfer (rejected once posted)
            tags:
                - documents
        get:
            operationId: getStockTransfer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StockTransfer'
                    description: Stock transfer
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
            summary: Get a stock transfer
            tags:
                - documents
        parameters:
            - $ref: '#/components/parameters/DocumentId'
        put:
            operationId: updateStockTransfer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StockTransferInput'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StockTransfer'
                    description: Draft updated
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
                "422":
                    $ref: '#/components/responses/ValidationError'
            summary: Replace a draft stock transfer (rejected once posted)
            tags:
                - documents
    /stock-transfers/{id}/post:
        parameters:
            - $ref: '#/components/parameters/DocumentId'
        post:
            operationId: postStockTransfer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StockTransfer'
                    description: Posted
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
            summary: Post a draft stock transfer (transfer_out + transfer_in pair)
            tags:
                - documents
    /stock-transfers/{id}/reverse:
        parameters:
            - $ref: '#/components/parameters/DocumentId'
        post:
            operationId: reverseStockTransfer
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StockTransfer'
                    description: Reversal document posted
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
            summary: Reverse a posted stock transfer (swaps the movement pair)
            tags:
                - documents
...
security:
    - sessionCookie: []
servers:
    - url: /v1